feat: replace old postprocess and validator with v2
