+++ validator/bundle.py
@@ -2,49 +2,16 @@ import json
 from collections.abc import Iterable, Mapping
 from dataclasses import (
     dataclass,
-    field,
 )
 from functools import cached_property
 from typing import (
     IO,
     Any,
     NotRequired,
-    Optional,
     TypedDict,
 )
 
 RESOURCE_REF = "/common-1.json#/definitions/resourceref"
-
-
-class InvalidBundleError(Exception):
-    pass
-
-
-@dataclass
-class GraphqlType:
-    type: str
-    spec: dict[str, Any]
-    bundle: "Bundle"
-
-    def __post_init__(self):  # noqa: D105
-        self.fields_by_name = {f.get("name"): f for f in self.spec.get("fields")}
-
-    def get_referenced_field_type(self, name: str) -> Optional["GraphqlType"]:
-        field = self.fields_by_name.get(name)
-        if field:
-            return self.bundle.get_graphql_type_by_name(field.get("type"))
-        return None
-
-    def get_interface_resolver_field(self) -> str | None:
-        return self.spec.get("interfaceResolve", {}).get("field")
-
-    def get_sub_type(self, discriminator: str) -> Optional["GraphqlType"]:
-        sub_type_name = (
-            self.spec.get("interfaceResolve", {}).get("fieldMap", {}).get(discriminator)
-        )
-        if sub_type_name:
-            return self.bundle.get_graphql_type_by_name(sub_type_name)
-        return None
 
 
 class InterfaceResolve(TypedDict):
@@ -62,7 +29,7 @@ class GraphqlField(TypedDict):
 
 
 @dataclass(frozen=True)
-class GraphqlTypeV2:
+class GraphqlType:
     name: str
     fields: dict[str, GraphqlField]
     datafile: str | None
@@ -145,13 +112,13 @@ class GraphqlLookup:
         )
 
     @staticmethod
-    def _build_graphql_type(conf: Mapping[str, Any]) -> GraphqlTypeV2:
+    def _build_graphql_type(conf: Mapping[str, Any]) -> GraphqlType:
         fields = {
             field_name: field
             for field in conf.get("fields") or []
             if (field_name := field.get("name"))
         }
-        return GraphqlTypeV2(
+        return GraphqlType(
             name=conf["name"],
             fields=fields,
             datafile=conf.get("datafile"),
@@ -162,7 +129,7 @@ class GraphqlLookup:
 
     @staticmethod
     def _build_top_level_graphql_type_names(
-        graphql_types: Iterable[GraphqlTypeV2],
+        graphql_types: Iterable[GraphqlType],
         graphql_type_names_with_schema: Iterable[str],
     ) -> set[str]:
         graphql_type_names = set(graphql_type_names_with_schema)
@@ -175,7 +142,7 @@ class GraphqlLookup:
 
     @staticmethod
     def _build_type_name_by_schema(
-        graphql_types: dict[str, GraphqlTypeV2],
+        graphql_types: dict[str, GraphqlType],
     ) -> dict[str, str]:
         type_name_by_type_datafile = {
             datafile: type_name
@@ -191,10 +158,10 @@ class GraphqlLookup:
             return type_name_by_type_datafile | type_name_by_query_datafile_schema
         return type_name_by_type_datafile
 
-    def get_by_type_name(self, name: str) -> GraphqlTypeV2 | None:
+    def get_by_type_name(self, name: str) -> GraphqlType | None:
         return self.graphql_types.get(name)
 
-    def get_by_schema(self, schema: str) -> GraphqlTypeV2 | None:
+    def get_by_schema(self, schema: str) -> GraphqlType | None:
         if name := self.type_name_by_schema.get(schema):
             return self.get_by_type_name(name)
         return None
@@ -209,72 +176,12 @@ class Bundle:
     git_commit: str
     git_commit_timestamp: str
 
-    _schema_to_graphql_type: dict[str, GraphqlType] = field(init=False)
-    _top_level_schemas: set[str] = field(init=False)
-    _graphql_type_by_name: dict[str, GraphqlType] = field(init=False)
-
     @cached_property
     def graphql_lookup(self) -> GraphqlLookup:
         if isinstance(self.graphql, dict):
             confs = self.graphql.get("confs", [])
             return GraphqlLookup(confs)
         return GraphqlLookup(self.graphql)
-
-    def __post_init__(self):  # noqa: D105
-        if isinstance(self.graphql, dict) and (self.graphql["confs"]):
-            self._graphql_type_by_name = {
-                t["name"]: GraphqlType(t["name"], t, self)
-                for t in self.graphql["confs"]
-            }
-        elif isinstance(self.graphql, list):
-            self._graphql_type_by_name = {
-                t["name"]: GraphqlType(t["name"], t, self) for t in self.graphql
-            }
-        else:
-            msg = (
-                "graphql field within bundle must be either "
-                "`list` or `dict` with keys `$schema` and `confs`"
-            )
-            raise InvalidBundleError(msg)
-        # use the datafile field on the graphql type to map to the schema
-        self._schema_to_graphql_type = {
-            f.get("datafile"): self._graphql_type_by_name[f["name"]]
-            for f in self.graphql.get("confs", [])
-            if f.get("datafile")
-        }
-        # also use the datafileSchema field within the Query section
-        self._schema_to_graphql_type.update({
-            f.get("datafileSchema"): self._graphql_type_by_name[f["type"]]
-            for f in self._graphql_type_by_name["Query"].spec["fields"]
-            if f.get("datafileSchema")
-        })
-        self._top_level_schemas = {
-            f.get("datafileSchema")
-            for f in self._graphql_type_by_name["Query"].spec["fields"]
-            if f.get("datafileSchema")
-        }
-
-    def to_dict(self):
-        return {
-            "git_commit": self.git_commit,
-            "git_commit_timestamp": self.git_commit_timestamp,
-            "schemas": self.schemas,
-            "graphql": self.graphql,
-            "data": self.data,
-            "resources": self.resources,
-        }
-
-    def get_graphql_type_for_schema(self, schema: str) -> GraphqlType | None:
-        return self._schema_to_graphql_type.get(schema)
-
-    def list_graphql_types(self) -> list[GraphqlType]:
-        return list(self._graphql_type_by_name.values())
-
-    def get_graphql_type_by_name(self, prop_type: str) -> GraphqlType | None:
-        return self._graphql_type_by_name.get(prop_type)
-
-    def is_top_level_schema(self, datafile_schema: str) -> bool:
-        return datafile_schema in self._top_level_schemas
 
 
 def load_bundle(bundle_source: IO) -> Bundle:
