--- validator/test/fixtures/get_schema_info_from_pointer/complex.yml
@@ -1,27 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schema:
-  type: object
-  properties:
-    a:
-      type: array
-      items:
-        properties:
-          b:
-            type: array
-            items:
-              properties:
-                c:
-                  type: object
-                  properties:
-                    d:
-                      type: object
-                      properties:
-                        e:
-                          <<: *magic
-
-ptr: /a/20/b/17/c/d/e
