--- validator/test/fixtures/get_schema_info_from_pointer/external_ref.yml
@@ -1,22 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schemas_bundle:
-  external_schema:
-      type: object
-      properties:
-        mirror:
-          <<: *magic
-
-schema:
-  type: object
-  properties:
-    internal:
-      type: array
-      items:
-        $ref: external_schema
-
-ptr: /internal/1/mirror
