--- validator/test/fixtures/get_schema_info_from_pointer/external_ref_obj.yml
@@ -1,24 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schemas_bundle:
-  external_schema:
-      type: object
-      properties:
-        mirror:
-          <<: *magic
-
-schema:
-  type: object
-  properties:
-    internal:
-      type: object
-      properties:
-        ext:
-          $ref: external_schema
-          desription: A description
-
-ptr: /internal/ext/mirror
