--- validator/test/fixtures/get_schema_info_from_pointer/external_ref_obj_oneof.yml
@@ -1,34 +1,0 @@
-magic_1: &magic_1
-  magic_object: is_here
-
-magic_2: &magic_2
-  another_magic_object: is_here
-
-result:
-- *magic_1
-- *magic_2
-
-schemas_bundle:
-  external_schema_1:
-    type: object
-    properties:
-      ext:
-        <<: *magic_1
-  external_schema_2:
-    type: object
-    properties:
-      ext:
-        <<: *magic_2
-
-schema:
-  type: object
-  properties:
-    internal:
-      type: object
-      properties:
-        mirror:
-          oneOf:
-            - $ref: external_schema_1
-            - $ref: external_schema_2
-
-ptr: /internal/mirror/ext
