--- validator/test/fixtures/get_schema_info_from_pointer/object.yml
@@ -1,13 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schema:
-  type: object
-  properties:
-    abc:
-      <<: *magic
-
-ptr: /abc
