--- validator/test/fixtures/get_schema_info_from_pointer/object_array.yml
@@ -1,15 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schema:
-  type: object
-  properties:
-    abc:
-      type: array
-      items:
-        <<: *magic
-
-ptr: /abc/0
