--- validator/test/fixtures/get_schema_info_from_pointer/object_object.yml
@@ -1,16 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schema:
-  type: object
-  properties:
-    abc:
-      type: object
-      properties:
-        efg:
-          <<: *magic
-
-ptr: /abc/efg
