--- validator/test/fixtures/get_schema_info_from_pointer/one_of.yml
@@ -1,26 +1,0 @@
-magic: &magic
-  magic_object: is_here
-
-result:
-- *magic
-
-schemas_bundle:
-  external_schema_1:
-    type: object
-    properties:
-      mirror:
-        <<: *magic
-  external_schema_2:
-    type: string
-
-schema:
-  type: object
-  properties:
-    internal:
-      type: array
-      items:
-        oneOf:
-          - $ref: external_schema_1
-          - $ref: external_schema_2
-
-ptr: /internal/1/mirror
