+++ validator/test/test_unique.py
@@ -3,7 +3,7 @@ from collections.abc import Callable
 import pytest
 
 from validator.bundle import Bundle
-from validator.postprocess_v2 import postprocess_bundle
+from validator.postprocess import postprocess_bundle
 
 
 @pytest.mark.parametrize(
@@ -69,4 +69,4 @@ def test_unique(
 
     postprocess_bundle(bundle)
 
-    assert bundle.to_dict() == expected_bundle.to_dict()
+    assert bundle == expected_bundle
